refactor: E2Eテストスクリプトの統合と簡素化
- run-e2e-parallel.tsとrun-e2e-with-db-optimized.tsを統合
- 新しいrun-e2e.tsが両方の機能を提供
- CI環境では自動的にワーカー数を1に制限
- ローカル環境では全ブラウザを並列実行（デフォルト）
- package.jsonのE2Eコマンドを整理して分かりやすく
- CI設定を新しいコマンド体系に合わせて更新
- ドキュメントも更新して最新の使い方を反映
