+++ .github/workflows/ci.yml
@@ -221,7 +221,7 @@ jobs:
           # Firefox専用のタイムアウト延長設定
           if [ "${{ matrix.browser }}" = "firefox" ]; then
             echo "🦊 Firefox detected: Using extended timeout (20 minutes)"
-            timeout 1200s npm run e2e:ci -- --project=${{ matrix.browser }} --workers=1 --reporter=dot || {
+            timeout 1200s npm run e2e:ci:${{ matrix.browser }} || {
               EXIT_CODE=$?
               echo "❌ Firefox E2E tests exited with code: $EXIT_CODE"
               if [ $EXIT_CODE -eq 124 ]; then
@@ -230,7 +230,7 @@ jobs:
               exit $EXIT_CODE
             }
           else
-            npm run e2e:ci -- --project=${{ matrix.browser }} --workers=1 --reporter=dot
+            npm run e2e:ci:${{ matrix.browser }}
           fi
 
       - name: Upload test report
