+++ docs/claude/commands.md
@@ -29,18 +29,13 @@ npx prisma generate      # Generate Prisma Client
 ### 🚀 Basic Execution
 
 ```bash
-npm run e2e              # Run all E2E tests with optimized database setup (recommended)
-npm run e2e:with-report  # Run tests and open HTML report
+npm run e2e              # Run all E2E tests (all browsers in parallel)
+npm run e2e:ui           # UI mode for debugging
+npm run e2e:debug        # Debug mode
+npm run e2e:report       # Show last test report
 ```
 
-### 🌐 Browser-specific Execution (for development only)
-
-```bash
-npm run e2e:chrome       # Chromium only (fastest)
-npm run e2e:cross-browser # Major browsers
-```
-
-### 📋 Feature-specific Execution (when needed)
+### 📋 Feature-specific Execution
 
 ```bash
 npm run e2e:smoke        # Smoke tests only
@@ -49,29 +44,48 @@ npm run e2e:materials    # Material management features
 npm run e2e:workflows    # Workflow tests
 ```
 
-### 🔍 Debugging & Detailed Execution
+### 🌐 Browser-specific Execution
 
 ```bash
-npm run e2e:ui       # UI mode for debugging
-npm run e2e:debug    # Debug mode
-npm run e2e:report   # Show last test report
+npm run e2e:chromium     # Chromium only
+npm run e2e:firefox      # Firefox only
+npm run e2e:webkit       # WebKit only
+```
 
+### 🏭 CI Execution
+
+```bash
+npm run e2e:ci:chromium  # CI mode for Chromium (workers=1)
+npm run e2e:ci:firefox   # CI mode for Firefox (workers=1)
+npm run e2e:ci:webkit    # CI mode for WebKit (workers=1)
+```
+
+### 🔍 Advanced Usage
+
+```bash
 # Run specific tests
 npm run e2e -- --grep "Equipment.*validation"
+
+# Run tests for specific browser with custom options
+npm run e2e -- --project=chromium --headed
+
+# Run tests with specific tags
+npm run e2e -- --grep "@critical"
 ```
 
 ### ⚠️ Notes
 
-- `npm run e2e` uses optimized database setup for 10x faster test execution
-- CI environment automatically limits workers to 1 for stability
-- Use feature-specific tests only when testing specific functionality
+- `npm run e2e` runs all browsers in parallel for comprehensive testing
+- Use browser-specific commands for faster feedback during development
+- CI commands automatically set workers=1 for stability
+- Use feature-specific tests to focus on specific functionality
 
-### Database Management
+### 📊 Database Management
 
 ```bash
-npm run e2e:db:create   # Create E2E test database
-npm run e2e:db:migrate  # Run migrations on E2E database
-npm run e2e:db:seed     # Seed E2E database with test data
-npm run e2e:db:drop     # Drop E2E test database
-npm run e2e:db:setup    # Complete setup (create + migrate + seed)
+npm run e2e:db:setup         # Complete setup (create + migrate + seed)
+npm run e2e:db:cleanup       # Clean up test database
+npm run e2e:db:cleanup-all   # Clean up all test databases
+npm run e2e:db:full-cleanup  # Full cleanup including templates
+npm run e2e:db:template-setup # Set up template database
 ```
