+++ package.json
@@ -16,32 +16,26 @@
     "worker:prod": "node dist/workers/index.js",
     "lint": "next lint",
     "test": "cross-env NODE_ENV=test jest",
-    "e2e": "tsx scripts/run-e2e-with-db-optimized.ts",
-    "e2e:parallel": "tsx scripts/run-e2e-parallel.ts",
-    "e2e:ci": "CI=true tsx scripts/run-e2e-with-db-optimized.ts",
-    "e2e:with-report": "tsx scripts/run-e2e-parallel.ts && playwright show-report",
-    "e2e:run": "playwright test",
+    "e2e": "tsx scripts/run-e2e.ts",
     "e2e:ui": "playwright test --ui",
     "e2e:debug": "playwright test --debug",
     "e2e:report": "playwright show-report",
+    "e2e:smoke": "npm run e2e -- --grep @smoke",
+    "e2e:master": "npm run e2e -- --grep @master",
+    "e2e:materials": "npm run e2e -- --grep @materials",
+    "e2e:workflows": "npm run e2e -- --grep @workflow",
+    "e2e:chromium": "npm run e2e -- --project=chromium",
+    "e2e:firefox": "npm run e2e -- --project=firefox",
+    "e2e:webkit": "npm run e2e -- --project=webkit",
+    "e2e:ci:chromium": "CI=true tsx scripts/run-e2e.ts -- --project=chromium",
+    "e2e:ci:firefox": "CI=true tsx scripts/run-e2e.ts -- --project=firefox",
+    "e2e:ci:webkit": "CI=true tsx scripts/run-e2e.ts -- --project=webkit",
     "e2e:db:setup": "tsx scripts/e2e-db-optimized.ts setup",
     "e2e:db:cleanup": "tsx scripts/e2e-db-optimized.ts cleanup",
     "e2e:db:cleanup-all": "tsx scripts/e2e-db-optimized.ts cleanup-all",
     "e2e:db:full-cleanup": "tsx scripts/e2e-db-optimized.ts full-cleanup",
     "e2e:db:template-setup": "tsx scripts/e2e-db-optimized.ts template-setup",
     "seed:test": "tsx scripts/seed-test-data.ts",
-    "e2e:smoke": "npm run e2e -- --grep @smoke --project=chromium",
-    "e2e:master": "npm run e2e -- --grep @master --project=chromium",
-    "e2e:materials": "npm run e2e -- --grep @materials --project=chromium",
-    "e2e:workflows": "npm run e2e -- --grep @workflow --project=chromium",
-    "e2e:chrome": "npm run e2e -- --project=chromium",
-    "e2e:firefox": "npm run e2e -- --project=firefox",
-    "e2e:firefox:ci": "CI=true FIREFOX_CI=true timeout 600s npm run e2e -- --project=firefox --workers=1 --retries=1",
-    "e2e:webkit": "npm run e2e -- --project=webkit",
-    "e2e:cross-browser": "npm run e2e -- --project=chromium --project=firefox",
-    "e2e:repeat": "tsx scripts/test-e2e-repeatedly.ts",
-    "e2e:benchmark": "tsx scripts/e2e-performance-benchmark.ts",
-    "e2e:compare": "tsx scripts/e2e-performance-comparison.ts",
     "prepare": "husky"
   },
   "overrides": {
